Make Navigation Bar responsive
Restyled the navigation button to be tilted the
same way as the rest of the page and added the
tilt to projects as well. Added the on hover
tilting to nav buttons and the project image
buttons as well. Added a hamburger menu that is
only to be used on smaller than 700px screens.
The hamburger functionality works and nav bar
is restyled with the same media query. The only
thing left in the functionality is to animate
the appearance of nav buttons.